requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.2
django-crispy-forms==2.0
django-import-export==3.2.0
django-paypal==2.0
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.4
import-export==0.3.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
phonenumbers==8.13.14
pytz==2023.3
PyYAML==6.0
requests==2.31.0
sqlparse==0.4.4
tablib==3.5.0
typing-extensions==4.6.3
urllib3==2.0.3
xlrd==2.0.1
xlwt==1.3.0
